requirements.txt
flask==2.0
requests==2.26.0
redis==4.1.0
python-dotenv==0.19.0
Werkzeug==2.0